// frontend/src/api/generated/models/partner-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PartnerDetail } from './partner-detail';

/**
 * 
 * @export
 * @interface PartnerPageResponse
 */
export interface PartnerPageResponse {
    /**
     * 
     * @type {Array<PartnerDetail>}
     * @memberof PartnerPageResponse
     */
    'content': Array<PartnerDetail>;
    /**
     * 
     * @type {number}
     * @memberof PartnerPageResponse
     */
    'page': number;
    /**
     * 
     * @type {number}
     * @memberof PartnerPageResponse
     */
    'size': number;
    /**
     * 
     * @type {number}
     * @memberof PartnerPageResponse
     */
    'totalElements': number;
    /**
     * 
     * @type {number}
     * @memberof PartnerPageResponse
     */
    'totalPages': number;
}